tests/wpt/web-platform-tests/fetch/api/headers/headers-basic.html

<!doctype html>
<html>
  <head>
    <meta charset="utf-8">
    <title>Headers structure</title>
    <meta name="help" href="https://fetch.spec.whatwg.org/#headers">
    <meta name="author" title="Canon Research France" href="https://www.crf.canon.fr">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <script>
      test(function() {
        new Headers();
      }, "Create headers from no parameter");

      test(function() {
        new Headers(undefined);
      }, "Create headers from undefined parameter");

      test(function() {
        new Headers({});
      }, "Create headers from empty object");

      var parameters = [null, 1];
      parameters.forEach(function(parameter) {
        test(function() {
          assert_throws(new TypeError(), () => new Headers(parameter));
        }, "Create headers with " + parameter + " should throw");
      });

      var headerDict = {"name1": "value1",
                        "name2": "value2",
                        "name3": "value3",
                        "name5": undefined,
                        "name4": null,
                        "Content-Type": "value4"
      };

      var headerSeq = [];
      for (var name in headerDict)
        headerSeq.push([name, headerDict[name]]);

      test(function() {
        var headers = new Headers(headerSeq);
        for (name in headerDict)
          assert_equals(headers.get(name), String(headerDict[name]),
            "name: " + name + " has value: " + headerDict[name]);
      }, "Create headers with sequence");

      test(function() {
        var headers = new Headers(headerDict);
        for (name in headerDict)
          assert_equals(headers.get(name), String(headerDict[name]),
            "name: " + name + " has value: " + headerDict[name]);
      }, "Create headers with OpenEndedDictionary");

      test(function() {
        var headers = new Headers(headerDict);
        var headers2 = new Headers(headers);
        for (name in headerDict)
          assert_equals(headers2.get(name), String(headerDict[name]),
            "name: " + name + " has value: " + headerDict[name]);
      }, "Create headers whith existing headers");

      test(function() {
        var headers = new Headers();
        for (name in headerDict) {
          headers.append(name, headerDict[name]);
          assert_equals(headers.get(name), String(headerDict[name]),
            "name: " + name + " has value: " + headerDict[name]);
        }
      }, "Check append method");

      test(function() {
        var headers = new Headers();
        for (name in headerDict) {
          headers.set(name, headerDict[name]);
          assert_equals(headers.get(name), String(headerDict[name]),
            "name: " + name + " has value: " + headerDict[name]);
        }
      }, "Check set method");

      test(function() {
        var headers = new Headers(headerDict);
        for (name in headerDict)
          assert_true(headers.has(name),"headers has name " + name);

        assert_false(headers.has("nameNotInHeaders"),"headers do not have header: nameNotInHeaders");
      }, "Check has method");

      test(function() {
        var headers = new Headers(headerDict);
        for (name in headerDict) {
          assert_true(headers.has(name),"headers have a header: " + name);
          headers.delete(name)
          assert_true(!headers.has(name),"headers do not have anymore a header: " + name);
        }
      }, "Check delete method");

      test(function() {
        var headers = new Headers(headerDict);
        for (name in headerDict)
          assert_equals(headers.get(name), String(headerDict[name]),
            "name: " + name + "has value: " + headerDict[name]);

        assert_equals(headers.get("nameNotInHeaders"), null, "header: nameNotInHeaders has no value");
      }, "Check get method");
    </script>
  </body>
</html>
