third_party/WebKit/LayoutTests/external/wpt/fetch/api/headers/headers-basic-expected.txt

This is a testharness.js-based test.
PASS Create headers from no parameter 
FAIL Create headers from undefined parameter Failed to construct 'Headers': No matching constructor signature.
PASS Create headers from empty object 
PASS Create headers with null should throw 
PASS Create headers with 1 should throw 
PASS Create headers with sequence 
PASS Create headers with record 
PASS Create headers with existing headers 
FAIL Create headers with existing headers with custom iterator assert_equals: expected (string) "test" but got (object) null
PASS Check append method 
PASS Check set method 
PASS Check has method 
PASS Check delete method 
PASS Check get method 
FAIL Check keys method assert_equals: expected object "[object Object]" but got object "[object Object]"
FAIL Check values method assert_equals: expected object "[object Object]" but got object "[object Object]"
FAIL Check entries method assert_equals: expected object "[object Object]" but got object "[object Object]"
PASS Check Symbol.iterator method 
PASS Check forEach method 
Harness: the test ran to completion.
